back-end/pyproject.toml
[tool.poetry]
name = "universal-disk-explorer"
version = "0.1.0"
description = "Universal Disk explorer with advanced file and video analysis capabilities"
authors = ["Haroon Abbasi <haroonabbasi.ali@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"  # Specify exact Python version range
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
aiofiles = "^23.1.0"
python-multipart = "^0.0.6"
ffmpeg-python = "^0.2.0"
Pillow = "^10.0.0"
imagehash = "^4.3.1"
python-magic = "^0.4.27"
pydantic = "^2.0.0"
pydantic-settings = "^2.7.1"
send2trash = "^1.8.0"  # Specify the version if needed

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.1"
black = "^23.7.0"
isort = "^5.12.0"
debugpy = "^1.8.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"